+++ CHANGES.md
@@ -16,7 +16,10 @@ Use list notation, and following prefixes:
 - Refactor: Change the get-size endpoint from https://services.virtusize.jp/ds-functions/size-rec/get-size ➝ https://size-recommendation.virtusize.jp/item
 - Refactor: Change in request body of VirtusizeGetSizeParams: Remove "modelInfo" key,  in "gender" now getting value from VirtusizeUserBodyProfile
 - Refactor: Change in response Model of BodyProfileRecommendedSize
+- Refactor: VirtusizeEnvironment enum cases key set to Uppercase 
+- Cleanup: Remove Print functions and unnecessary commented code from BodyProfileRecommendedSize, VirtusizeAPIService, VirtusizeAPIRequest
 - Docs: Update Readme file to update pod version 
+
 
 
 ### 2.5.4